+++ OpenSim/Analyses/StaticOptimization.h
@@ -124,6 +124,7 @@ public:
     void setStorageCapacityIncrements(int) {}
     Storage* getActivationStorage();
     Storage* getForceStorage();
+    Storage* getStatesDerivativeStore() {return _statesDerivativeStore; }
 
     bool getUseModelForceSet() { return _useModelForceSet; }
     void setUseModelForceSet(bool aUseModelActuatorSet) { _useModelForceSet = aUseModelActuatorSet; }
@@ -137,7 +138,7 @@ public:
     double getConvergenceCriterion() { return _convergenceCriterion; }
     void setMaxIterations( const int maxIt) { _maximumIterations = maxIt; }
     int getMaxIterations() {return _maximumIterations; }
-    Storage* getStatesDerivativeStore() {return _statesDerivativeStore; }
+    
     //--------------------------------------------------------------------------
     // ANALYSIS
     //--------------------------------------------------------------------------